src/questoespmp2/requirements.txt
kivy>=2.1.0
kivymd>=1.1.1
openai>=0.27.0
python-dotenv>=0.20.0
easygui>=0.98.3
numpy>=1.22.0
pandas>=1.4.0
scikit-learn>=1.0.0
nltk>=3.7.0 